requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.25.0
feedparser>=6.0.11
beautifulsoup4>=4.12.0
lxml>=5.1.0
schedule>=1.2.1
